refactor: Section 도메인 api 에 노출되지 않도록 변경
